added cocoapods. some project organization
