--- CalorieTracker/CalorieTracker/ViewController.swift
@@ -1,20 +1,0 @@
-//
-//  ViewController.swift
-//  CalorieTracker
-//
-//  Created by Joe Veverka on 6/19/20.
-//  Copyright © 2020 Joe Veverka. All rights reserved.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
