+++ CalorieTracker/Podfile
@@ -1,0 +1,12 @@
+# Uncomment the next line to define a global platform for your project
+# platform :ios, '9.0'
+
+target 'CalorieTracker' do
+  # Comment the next line if you don't want to use dynamic frameworks
+  use_frameworks!
+
+  # Pods for CalorieTracker
+	pod "SwiftChart", :git => 'https://github.com/gpbl/SwiftChart', :branch => 'master'
+	pod 'SwiftLint'
+
+end
